as expected, dqn isn´t working
even for a one time rotated cube the dqn isn´t able to easily solve every state. But dqn  wan´t supposed so solve the problems. It´s purpose was to test and build the framework. 

Now that the framework is build, I´ll try to rebuild the model from this paper to solve the combinatorical problem: https://arxiv.org/pdf/1805.07470.pdf